fix numeric

+++ polynomial/polynomial.class.rb
@@ -57,7 +57,7 @@ class Polynomial
   end
   
   def calc x
-    Error.call "Polynomial::calc: x is not an integer" if !x.is_a?Integer
+    Error.call "Polynomial::calc: x is not an integer" if !x.is_a?Numeric
     
     y = 0
     @coef.each do |coef,value|
